+++ CHANGELOG.md
@@ -1,5 +1,29 @@
 # Change Log
 
+## 2017-01-22 Chase Hennion
+
+* Adding work that covers lectures 48 and 49, which is the end of Section 2.
+
+## 2017-01-21 Chase Hennion
+
+* Adding work that covers lectures 46 through 48 in Section 2.
+* Adding isograms.txt, a file that contains a variety of 3, 4, 5, 6 and 7 letter isograms.
+* Adding handling to game to ask player for desired isogram length
+* Adding handling to game to read isograms from isograms.txt and select one at random for each game based upon the desired isogram length.
+* Adding a neat ASCII banner to the intro.
+
+## 2017-01-20 Chase Hennion
+
+* Adding work that covers lectures 37 through 45 in Section 2.
+
+## 2017-01-19 Chase Hennion
+
+* Adding work that covers lectures 31 through 36 in Section 2.
+
+## 2017-01-17 Chase Hennion
+
+* Adding work that covers lectures 21 through 30 in Section 2.
+
 ## 2017-01-16 Chase Hennion
 
 * Adding initital work